Add Parse(Stream) and ParseAsync(Stream)
