+++ src/System.Text.Json/src/System.Text.Json.csproj
@@ -16,6 +16,7 @@
     <Compile Include="System\Text\Json\JsonDocument.CustomDb.cs" />
     <Compile Include="System\Text\Json\JsonDocument.CustomStack.cs" />
     <Compile Include="System\Text\Json\JsonDocument.DbRow.cs" />
+    <Compile Include="System\Text\Json\JsonDocument.Parse.cs" />
     <Compile Include="System\Text\Json\JsonDocument.StackRow.cs" />
     <Compile Include="System\Text\Json\JsonDocument.TryGetProperty.cs" />
     <Compile Include="System\Text\Json\JsonElement.cs" />
