+++ src/System.Text.Json/src/System/Text/Json/JsonDocument.cs
@@ -47,104 +47,6 @@ namespace System.Text.Json
             }
         }
 
-        public static JsonDocument Parse(ReadOnlyMemory<byte> utf8Json, JsonReaderOptions readerOptions = default)
-        {
-            CheckSupportedOptions(readerOptions);
-
-            return Parse(utf8Json, readerOptions, null);
-        }
-
-        public static JsonDocument Parse(ReadOnlySequence<byte> utf8Json, JsonReaderOptions readerOptions = default)
-        {
-            CheckSupportedOptions(readerOptions);
-
-            if (utf8Json.IsSingleSegment)
-            {
-                return Parse(utf8Json.First, readerOptions, null);
-            }
-
-            int length = checked((int)utf8Json.Length);
-            byte[] utf8Bytes = ArrayPool<byte>.Shared.Rent(length);
-
-            try
-            {
-                utf8Bytes.CopyTo(utf8Bytes.AsSpan());
-                return Parse(utf8Bytes.AsMemory(0, length), readerOptions, utf8Bytes);
-            }
-            catch
-            {
-                utf8Bytes.AsSpan(0, length).Clear();
-                ArrayPool<byte>.Shared.Return(utf8Bytes);
-                throw;
-            }
-        }
-
-        public static JsonDocument Parse(ReadOnlyMemory<char> json, JsonReaderOptions readerOptions = default)
-        {
-            CheckSupportedOptions(readerOptions);
-
-            ReadOnlySpan<char> jsonChars = json.Span;
-            int byteCount = Utf8JsonReader.Utf8Encoding.GetByteCount(jsonChars);
-            byte[] utf8Bytes = ArrayPool<byte>.Shared.Rent(byteCount);
-
-            try
-            {
-                int byteCount2 = Utf8JsonReader.Utf8Encoding.GetBytes(jsonChars, utf8Bytes);
-                Debug.Assert(byteCount == byteCount2);
-
-                return Parse(utf8Bytes.AsMemory(0, byteCount2), readerOptions, utf8Bytes);
-            }
-            catch
-            {
-                utf8Bytes.AsSpan(0, byteCount).Clear();
-                ArrayPool<byte>.Shared.Return(utf8Bytes);
-                throw;
-            }
-        }
-
-        public static JsonDocument Parse(string json, JsonReaderOptions readerOptions = default)
-        {
-            CheckSupportedOptions(readerOptions);
-
-            return Parse(json.AsMemory(), readerOptions);
-        }
-
-        private static JsonDocument Parse(
-            ReadOnlyMemory<byte> utf8Json,
-            JsonReaderOptions readerOptions,
-            byte[] extraRentedBytes)
-        {
-            ReadOnlySpan<byte> utf8JsonSpan = utf8Json.Span;
-            Utf8JsonReader reader = new Utf8JsonReader(
-                utf8JsonSpan,
-                isFinalBlock: true,
-                new JsonReaderState(JsonReaderState.DefaultMaxDepth, readerOptions));
-
-            var database = new CustomDb(utf8Json.Length);
-            var stack = new CustomStack(JsonReaderState.DefaultMaxDepth * StackRow.Size);
-
-            try
-            {
-                Parse(utf8JsonSpan, reader, ref database, ref stack);
-
-                if (database.Length == 0)
-                {
-                    throw new JsonReaderException("Cannot load the empty document", -1, -1);
-                }
-            }
-            catch
-            {
-                database.Dispose();
-                throw;
-            }
-            finally
-            {
-                stack.Dispose();
-            }
-
-            return new JsonDocument(utf8Json, database, extraRentedBytes);
-        }
-
         internal JsonTokenType GetJsonTokenType(int index)
         {
             CheckNotDisposed();
